move ProxyClient, ProxyConfiguration and related structs to mod.rs
