+++ lib/src/network/mod.rs
@@ -1,11 +1,16 @@
 #![allow(dead_code, unused_must_use, unused_variables, unused_imports)]
 
 use mio;
-use mio::Ready;
+use mio::{Poll,Ready};
 use mio::unix::UnixReady;
 use std::fmt;
 use std::net::SocketAddr;
+use std::rc::Rc;
+use std::cell::RefCell;
+use slab::{Entry,VacantEntry};
 use time::{precise_time_ns,SteadyTime,Duration};
+
+use sozu_command::messages::{OrderMessage,OrderMessageAnswer};
 
 pub mod buffer_queue;
 #[macro_use] pub mod metrics;
@@ -56,6 +61,45 @@ pub enum Protocol {
   HTTP,
   HTTPS,
   TCP
+}
+
+pub trait ProxyClient {
+  fn protocol(&self)  -> Protocol;
+  fn ready(&mut self) -> ClientResult;
+  fn process_events(&mut self, token: Token, events: Ready);
+  fn close(&mut self, poll: &mut Poll, configuration: &mut ProxyConfiguration<Self>) -> Vec<Token>;
+  fn close_backend(&mut self, token: Token, poll: &mut Poll, configuration: &mut ProxyConfiguration<Self>);
+}
+
+#[derive(Clone,Copy,Debug,PartialEq)]
+pub enum BackendConnectionStatus {
+  NotConnected,
+  Connecting,
+  Connected,
+}
+
+#[derive(Debug,PartialEq)]
+pub enum BackendConnectAction {
+  New,
+  Reuse,
+  Replace,
+}
+
+#[derive(Debug,PartialEq)]
+pub enum AcceptError {
+  IoError,
+  TooManyClients,
+  WouldBlock,
+}
+
+use self::session::{ClientToken,ListenToken};
+pub trait ProxyConfiguration<Client> {
+  fn connect_to_backend(&mut self, event_loop: &mut Poll, client: Rc<RefCell<Client>>,
+    entry: Entry<Rc<RefCell<Client>>, ClientToken>, back_token: Token) ->Result<BackendConnectAction,ConnectionError>;
+  fn notify(&mut self, event_loop: &mut Poll, message: OrderMessage) -> OrderMessageAnswer;
+  fn accept(&mut self, token: ListenToken, event_loop: &mut Poll, entry: VacantEntry<Rc<RefCell<Client>>, ClientToken>,
+           client_token: Token) -> Result<(ClientToken, bool), AcceptError>;
+  fn close_backend(&mut self, app_id: String, addr: &SocketAddr);
 }
 
 #[derive(Debug,PartialEq,Eq)]
@@ -158,7 +202,7 @@ pub enum ConnectionError {
 #[derive(Debug,PartialEq,Eq)]
 pub enum SocketType {
   Listener,
-  FrontClient,
+  FrontClient
 }
 
 #[derive(Debug,PartialEq,Eq)]
