+++ lib/src/network/session.rs
@@ -24,8 +24,8 @@ use sozu_command::channel::Channel;
 use sozu_command::messages::{self,TcpFront,Order,Instance,MessageId,OrderMessage,
   OrderMessageAnswer,OrderMessageStatus};
 
-use network::{ClientResult,ConnectionError,
-  SocketType,Protocol,RequiredEvents};
+use network::{ClientResult,ConnectionError,SocketType,Protocol,RequiredEvents,
+  ProxyClient,ProxyConfiguration,AcceptError,BackendConnectAction};
 
 const SERVER: Token = Token(0);
 const DEFAULT_FRONT_TIMEOUT: u64 = 50000;
@@ -60,44 +60,6 @@ impl From<ClientToken> for usize {
     fn from(val: ClientToken) -> usize {
         val.0
     }
-}
-
-pub trait ProxyClient {
-  fn protocol(&self)  -> Protocol;
-  fn ready(&mut self) -> ClientResult;
-  fn process_events(&mut self, token: Token, events: Ready);
-  fn close(&mut self, poll: &mut Poll, configuration: &mut ProxyConfiguration<Self>) -> Vec<Token>;
-  fn close_backend(&mut self, token: Token, poll: &mut Poll, configuration: &mut ProxyConfiguration<Self>);
-}
-
-#[derive(Clone,Copy,Debug,PartialEq)]
-pub enum BackendConnectionStatus {
-  NotConnected,
-  Connecting,
-  Connected,
-}
-
-#[derive(Debug,PartialEq)]
-pub enum BackendConnectAction {
-  New,
-  Reuse,
-  Replace,
-}
-
-#[derive(Debug,PartialEq)]
-pub enum AcceptError {
-  IoError,
-  TooManyClients,
-  WouldBlock,
-}
-
-pub trait ProxyConfiguration<Client> {
-  fn connect_to_backend(&mut self, event_loop: &mut Poll, client: Rc<RefCell<Client>>,
-    entry: Entry<Rc<RefCell<Client>>, ClientToken>, back_token: Token) ->Result<BackendConnectAction,ConnectionError>;
-  fn notify(&mut self, event_loop: &mut Poll, message: OrderMessage) -> OrderMessageAnswer;
-  fn accept(&mut self, token: ListenToken, event_loop: &mut Poll, entry: VacantEntry<Rc<RefCell<Client>>, ClientToken>,
-           client_token: Token) -> Result<(ClientToken, bool), AcceptError>;
-  fn close_backend(&mut self, app_id: String, addr: &SocketAddr);
 }
 
 pub struct Session<ServerConfiguration,Client> {
